Initial commit of a non-underhanded version of the 2015 underhanded c competition.
In this version, we use the Pearson Correlation of the test and reference arrays as our
similarity metric. From here, the current idea is that if we can subtly introduce a bug
that for a certain input test array causes a non-zero covariance while simultaneously
causing a zero standard deviation for the same input, then we would effect an (-)inf
result set. Since a non-buggy implementation of the Pearson Correlation would cause the
covariance of test to equal zero if its standard deviation did as well, we correctly check for
nan results and return 0 accordingly. However we do not check for (-)inf's therefore they
get compared against the threshold to determine if we return 0 or 1. Therefore, with such
a bug, we can simply set the covariance to be inf by having the median be on average greater
than the the deviation in our test, bug-tickling, input.

That's the idea, anyway. My current thought is to try and set up a simple rounding error in
the standard deviation function such that e.g. [1.0, 1.000001, 1.0003] would result in [1, 1, 1].
We'll see how non-obvious i can make that... Might end up not being the route to go.